# swagger_server/models/get_vp_response_object.py
# coding: utf-8

from __future__ import absolute_import
from swagger_server.models.voyage_plan import VoyagePlan
from .base_model_ import Model
from datetime import date, datetime
from typing import List, Dict
from ..util import deserialize_model


class GetVPResponseObject(Model):
    """
    NOTE: This class is auto generated by the swagger code generator program.
    Do not edit the class manually.
    """
    def __init__(self, last_interaction_time: datetime=None, subject: str=None, voyage_plan: VoyagePlan=None):
        """
        GetVPResponseObject - a model defined in Swagger

        :param last_interaction_time: The last_interaction_time of this GetVPResponseObject.
        :type last_interaction_time: datetime
        :param subject: The subject of this GetVPResponseObject.
        :type subject: str
        :param voyage_plan: The voyage_plan of this GetVPResponseObject.
        :type voyage_plan: VoyagePlan
        """
        self.swagger_types = {
            'last_interaction_time': datetime,
            'subject': str,
            'voyage_plan': VoyagePlan
        }

        self.attribute_map = {
            'last_interaction_time': 'lastInteractionTime',
            'subject': 'subject',
            'voyage_plan': 'voyagePlan'
        }

        self._last_interaction_time = last_interaction_time
        self._subject = subject
        self._voyage_plan = voyage_plan

    @classmethod
    def from_dict(cls, dikt) -> 'GetVPResponseObject':
        """
        Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The GetVPResponseObject of this GetVPResponseObject.
        :rtype: GetVPResponseObject
        """
        return deserialize_model(dikt, cls)

    @property
    def last_interaction_time(self) -> datetime:
        """
        Gets the last_interaction_time of this GetVPResponseObject.

        :return: The last_interaction_time of this GetVPResponseObject.
        :rtype: datetime
        """
        return self._last_interaction_time

    @last_interaction_time.setter
    def last_interaction_time(self, last_interaction_time: datetime):
        """
        Sets the last_interaction_time of this GetVPResponseObject.

        :param last_interaction_time: The last_interaction_time of this GetVPResponseObject.
        :type last_interaction_time: datetime
        """
        if last_interaction_time is None:
            raise ValueError("Invalid value for `last_interaction_time`, must not be `None`")

        self._last_interaction_time = last_interaction_time

    @property
    def subject(self) -> str:
        """
        Gets the subject of this GetVPResponseObject.

        :return: The subject of this GetVPResponseObject.
        :rtype: str
        """
        return self._subject

    @subject.setter
    def subject(self, subject: str):
        """
        Sets the subject of this GetVPResponseObject.

        :param subject: The subject of this GetVPResponseObject.
        :type subject: str
        """
        if subject is None:
            raise ValueError("Invalid value for `subject`, must not be `None`")

        self._subject = subject

    @property
    def voyage_plan(self) -> VoyagePlan:
        """
        Gets the voyage_plan of this GetVPResponseObject.

        :return: The voyage_plan of this GetVPResponseObject.
        :rtype: VoyagePlan
        """
        return self._voyage_plan

    @voyage_plan.setter
    def voyage_plan(self, voyage_plan: VoyagePlan):
        """
        Sets the voyage_plan of this GetVPResponseObject.

        :param voyage_plan: The voyage_plan of this GetVPResponseObject.
        :type voyage_plan: VoyagePlan
        """

        self._voyage_plan = voyage_plan