# swagger_server/models/latitude_type.py
# coding: utf-8

from __future__ import absolute_import
from .base_model_ import Model
from datetime import date, datetime
from typing import List, Dict
from ..util import deserialize_model


class LatitudeType(Model):
    """
    NOTE: This class is auto generated by the swagger code generator program.
    Do not edit the class manually.
    """
    def __init__(self):
        """
        LatitudeType - a model defined in Swagger

        """
        self.swagger_types = {
            
        }

        self.attribute_map = {
            
        }


    @classmethod
    def from_dict(cls, dikt) -> 'LatitudeType':
        """
        Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The LatitudeType of this LatitudeType.
        :rtype: LatitudeType
        """
        return deserialize_model(dikt, cls)