# swagger_server/models/longitude_type.py
# coding: utf-8

from __future__ import absolute_import
from .base_model_ import Model
from datetime import date, datetime
from typing import List, Dict
from ..util import deserialize_model


class LongitudeType(Model):
    """
    NOTE: This class is auto generated by the swagger code generator program.
    Do not edit the class manually.
    """
    def __init__(self):
        """
        LongitudeType - a model defined in Swagger

        """
        self.swagger_types = {
            
        }

        self.attribute_map = {
            
        }


    @classmethod
    def from_dict(cls, dikt) -> 'LongitudeType':
        """
        Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The LongitudeType of this LongitudeType.
        :rtype: LongitudeType
        """
        return deserialize_model(dikt, cls)